+++ @navikt/aksel/src/codemod/transforms/darkside/token-update/tests/complete.input.js
@@ -1,0 +1,10 @@
+export const testComponent = () => {
+  return (
+    <div
+      className="test bg-(--a-surface-neutral)"
+      style={{ backgroundColor: "var(--a-surface-subtle)" }}
+    >
+      Test
+    </div>
+  );
+};