+++ @navikt/core/react/src/form/file-upload/parts/item/ItemButton.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { ArrowsCirclepathIcon, TrashIcon } from "@navikt/aksel-icons";
 import { Button } from "../../../../button";
+import { useRenameCSS } from "../../../../theme/Theme";
 
 interface Props {
   action: "delete" | "retry";
@@ -10,12 +11,13 @@ interface Props {
 }
 
 const ItemButton = ({ action, onClick, id, title }: Props) => {
+  const { cn } = useRenameCSS();
   const Icon = action === "delete" ? TrashIcon : ArrowsCirclepathIcon;
 
   return (
     <Button
       id={id}
-      className="navds-file-item__button"
+      className={cn("navds-file-item__button")}
       type="button"
       variant="tertiary-neutral"
       onClick={onClick}