+++ @navikt/core/react/src/tooltip/Tooltip.tsx
@@ -11,12 +11,11 @@ import {
   useHover,
   useInteractions,
 } from "@floating-ui/react";
-import cl from "clsx";
 import React, { HTMLAttributes, forwardRef, useRef } from "react";
 import { useModalContext } from "../modal/Modal.context";
 import { Portal } from "../portal";
 import { Slot } from "../slot/Slot";
-import { useThemeInternal } from "../theme/Theme";
+import { useRenameCSS } from "../theme/Theme";
 import { Detail } from "../typography";
 import { useId } from "../util/hooks";
 import { useControllableState } from "../util/hooks/useControllableState";
@@ -124,8 +123,7 @@ export const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
     },
     ref,
   ) => {
-    const themeContext = useThemeInternal(false);
-    const showArrow = _arrow && !themeContext;
+    const { cn } = useRenameCSS();
 
     const [_open, _setOpen] = useControllableState({
       defaultValue: defaultOpen,
@@ -153,7 +151,7 @@ export const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
       open: _open,
       onOpenChange: (newState) => _setOpen(newState),
       middleware: [
-        offset(_offset ?? (themeContext ? 8 : _arrow ? 16 : 4)),
+        offset(_offset ?? (_arrow ? 8 : 4)),
         shift(),
         flip({ padding: 5, fallbackPlacements: ["bottom", "top"] }),
         flArrow({ element: arrowRef, padding: 5 }),
@@ -225,7 +223,7 @@ export const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
                 },
                 role: "tooltip",
                 id: ariaId,
-                className: cl(
+                className: cn(
                   "navds-tooltip",
                   "navds-detail navds-detail--small",
                   className,
@@ -236,20 +234,24 @@ export const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
             >
               {content}
               {keys && (
-                <span className="navds-tooltip__keys" aria-hidden>
+                <span className={cn("navds-tooltip__keys")} aria-hidden>
                   {keys.map((key) => (
-                    <Detail as="kbd" key={key} className="navds-tooltip__key">
+                    <Detail
+                      as="kbd"
+                      key={key}
+                      className={cn("navds-tooltip__key")}
+                    >
                       {key}
                     </Detail>
                   ))}
                 </span>
               )}
-              {showArrow && (
+              {_arrow && (
                 <div
                   ref={(node) => {
                     arrowRef.current = node;
                   }}
-                  className="navds-tooltip__arrow"
+                  className={cn("navds-tooltip__arrow")}
                   style={{
                     left: arrowX != null ? `${arrowX}px` : "",
                     top: arrowY != null ? `${arrowY}px` : "",