+++ aksel.nav.no/website/e2e/search.e2e.ts
@@ -3,8 +3,8 @@ import { expect, test } from "@playwright/test";
 test.describe("Check website search", () => {
   test("Check newest article list", async ({ page }) => {
     await page.goto("http://localhost:3000/");
-    await page.waitForLoadState("networkidle");
-    await page.waitForTimeout(3000);
+    await page.waitForLoadState("domcontentloaded");
+    await page.waitForTimeout(5000);
     await page.getByRole("button", { name: "Søk" }).click();
 
     /* Check that we have more than 0 "nyeste artikler" */
@@ -17,8 +17,8 @@ test.describe("Check website search", () => {
 
   test("Test searching for 'link'", async ({ page }) => {
     await page.goto("http://localhost:3000/");
-    await page.waitForLoadState("networkidle");
-    await page.waitForTimeout(3000);
+    await page.waitForLoadState("domcontentloaded");
+    await page.waitForTimeout(5000);
     await page.getByRole("button", { name: "Søk" }).click();
 
     await page.getByPlaceholder("Søk på artikler, f.eks. Button").fill("link");