+++ aksel.nav.no/website/pages/_app.tsx
@@ -1,50 +1,44 @@
 import { AppProps } from "next/app";
 import { useEffect } from "react";
-import { useCheckAuth } from "@/hooks/useCheckAuth";
+import "@navikt/ds-tokens/darkside-css";
 import { useHashScroll } from "@/hooks/useHashScroll";
-import { SanityDataContext } from "@/hooks/useSanityData";
+import { ConsentBanner } from "@/web/ConsentBanner";
+import { CookieProvider } from "@/web/CookieProvider";
+import { SanityDataProvider } from "@/web/SanityDataProvider";
+import { Umami } from "@/web/Umami";
 import { BaseSEO } from "@/web/seo/BaseSEO";
 import "../components/styles/index.css";
 
 function App({ Component, pageProps, router }: AppProps) {
   useHashScroll();
 
-  /* As of 01.01.25, removed until cookie compliance is implemented */
-  /* useAmplitudeInit(); */
-
   useEffect(() => {
     window.location.host === "design.nav.no" &&
       window.location.replace(`http://aksel.nav.no`);
-
-    /**
-     * Midlertidig utkommentert for å unngå lasting av hotjar-bundle
-     * Package.json: "react-hotjar": "^6.1.0",
-     * Import: import { hotjar } from "react-hotjar";
-     * Script: hotjar.initialize(148751, 6);
-     */
   }, []);
 
   const useGlobalStyles =
     !router.pathname.startsWith("/templates/") &&
     !router.pathname.startsWith("/eksempler/");
 
-  const validUser = useCheckAuth(!useGlobalStyles);
-
   return (
-    <>
+    <CookieProvider>
       <BaseSEO path={router.asPath} />
-      <SanityDataContext.Provider
-        value={{ id: pageProps?.id ?? pageProps?.page?._id, validUser }}
-      >
-        {useGlobalStyles ? (
+      <Umami />
+      <ConsentBanner
+        hide={!useGlobalStyles}
+        defaultShow={pageProps.showCookieBanner}
+      />
+      {useGlobalStyles ? (
+        <SanityDataProvider id={pageProps?.id ?? pageProps?.page?._id}>
           <div className="globalstyles">
             <Component {...pageProps} />
           </div>
-        ) : (
-          <Component {...pageProps} />
-        )}
-      </SanityDataContext.Provider>
-    </>
+        </SanityDataProvider>
+      ) : (
+        <Component {...pageProps} />
+      )}
+    </CookieProvider>
   );
 }